back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.20.0",
    "dotenv>=0.9.9",
    "fastapi>=0.124.2",
    "openai-agents>=0.6.2",
    "qdrant-client>=1.16.1",
    "trafilatura>=2.0.0",
    "better-auth>=0.2.2",
    "sqlalchemy>=2.0.0",
]
